gcp_vertex/requirements.txt
pyyaml == 6.0
google-cloud-storage == 2.3.0
google-cloud-aiplatform == 1.13.1
torch
